requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
pydantic
redis
python-dotenv
